requirements.txt
pywin32
pillow
rich
comtypes
